# openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TimingApp.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/api/passages": {
      "post": {
        "tags": [
          "Control"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PassageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassageReceipt"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/control/start": {
      "post": {
        "tags": [
          "Control"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinishRecordingStatus"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/control/trigger": {
      "put": {
        "tags": [
          "Control"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinishRecordingStatus"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CameraState": {
        "enum": [
          "Off",
          "Disabled",
          "Starting",
          "Running",
          "Error"
        ]
      },
      "CameraStatus": {
        "required": [
          "state",
          "errorCode",
          "measuredFrameRate",
          "configuredFrameRate"
        ],
        "type": "object",
        "properties": {
          "state": {
            "$ref": "#/components/schemas/CameraState"
          },
          "errorCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "measuredFrameRate": {
            "type": "number",
            "format": "double"
          },
          "configuredFrameRate": {
            "type": "integer",
            "format": "int32"
          },
          "warningCode": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "FinishRecordingStatus": {
        "required": [
          "mode",
          "finishCamera",
          "frontCamera",
          "line",
          "overlay",
          "pendingSaves",
          "lastProblem",
          "serverTime"
        ],
        "type": "object",
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/OperatingMode"
          },
          "finishCamera": {
            "$ref": "#/components/schemas/CameraStatus"
          },
          "frontCamera": {
            "$ref": "#/components/schemas/CameraStatus"
          },
          "line": {
            "$ref": "#/components/schemas/LineStatus"
          },
          "overlay": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LineOverlay"
              }
            ]
          },
          "pendingSaves": {
            "type": "integer",
            "format": "int32"
          },
          "lastProblem": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SaveProblem"
              }
            ]
          },
          "serverTime": {
            "type": "string",
            "format": "date-time"
          },
          "raceName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "LineOverlay": {
        "required": [
          "position",
          "width"
        ],
        "type": "object",
        "properties": {
          "position": {
            "type": "number",
            "format": "double"
          },
          "width": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "LineState": {
        "enum": [
          "Free",
          "Occupied",
          "Recording"
        ]
      },
      "LineStatus": {
        "required": [
          "state",
          "occupancyPercent",
          "lineRate",
          "lineRateWarning",
          "manualTrigger",
          "backgroundLearning",
          "eventRunning",
          "eventStartedAt",
          "detectedEvents"
        ],
        "type": "object",
        "properties": {
          "state": {
            "$ref": "#/components/schemas/LineState"
          },
          "occupancyPercent": {
            "type": "number",
            "format": "double"
          },
          "lineRate": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "lineRateWarning": {
            "type": "boolean"
          },
          "manualTrigger": {
            "type": "boolean"
          },
          "backgroundLearning": {
            "type": "boolean"
          },
          "eventRunning": {
            "type": "boolean"
          },
          "eventStartedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "detectedEvents": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "OperatingMode": {
        "enum": [
          "Stopped",
          "Preview",
          "Recording"
        ]
      },
...
      "PassageReceipt": {
        "required": [
          "startNumber",
          "time"
        ],
        "type": "object",
        "properties": {
          "startNumber": {
            "type": "string"
          },
          "time": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PassageRequest": {
        "required": [
          "startNumber",
          "time"
        ],
        "type": "object",
        "properties": {
          "startNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "time": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SaveProblem": {
        "required": [
          "code",
          "recordingId",
          "at"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "recordingId": {
            "type": [
              "null",
              "string"
            ]
          },
          "at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "StartRequest": {
        "required": [
          "mode"
        ],
        "type": "object",
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/OperatingMode"
          },
          "raceName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "TriggerRequest": {
        "required": [
          "active"
        ],
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}